requirements.txt
numpy=1.18.5
scipy=1.4.1
Pillow=8.0.1
cython=0.29.21
matplotlib=3.3.3
scikit-image=0.16.2
tensorflow>=2.3.0
keras>=2.4.0
opencv-python=4.4.0
h5py=2.10.0
imgaug=0.4.0
IPython[all]